+++ cognite/client/data_classes/simulators/__init__.py
@@ -1,1 +1,23 @@
 from __future__ import annotations
+
+from cognite.client.data_classes.simulators.simulators import (
+    Simulator,
+    SimulatorIntegration,
+    SimulatorIntegrationList,
+    SimulatorList,
+    SimulatorStep,
+    SimulatorStepField,
+    SimulatorStepOption,
+    SimulatorUnitEntry,
+)
+
+__all__ = [
+    "Simulator",
+    "SimulatorIntegration",
+    "SimulatorIntegrationList",
+    "SimulatorList",
+    "SimulatorStep",
+    "SimulatorStepField",
+    "SimulatorStepOption",
+    "SimulatorUnitEntry",
+]