+++ tests/tests_unit/test_api_client.py
@@ -1366,6 +1366,9 @@ class TestRetryableEndpoints:
                 ("POST", "https://api.cognitedata.com/api/v1/projects/bla/context/diagram/detect", True),
                 ("GET", "https://api.cognitedata.com/api/v1/projects/bla/context/diagram/convert/123", True),
                 ("GET", "https://api.cognitedata.com/api/v1/projects/bla/context/diagram/detect/456", True),
+                # Simulators
+                ("POST", "https://api.cognitedata.com/api/v1/projects/bla/simulators/list", True),
+                ("POST", "https://api.cognitedata.com/api/v1/projects/bla/simulators/delete", False),
             ]
         ),
     )